requirements.txt
Django==1.9.4
django-extensions==1.6.1
djangorestframework==3.3.2
six==1.10.0
Werkzeug==0.11.4
